se arregla home con menu lateral
