+++ src/routes/Routes.jsx
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
 import {BrowserRouter as Router, Route, Redirect } from 'react-router-dom';
-import Home from "../controllers/home/HomeV2";
+import Home from "../controllers/home/Home";
 import Login from "../controllers/login/Login";
 import {PrivateRoute} from "./PrivateRoute";
 import tokenChecker from '../utils/TokenChecker';
